feat(answer): #FOR-873 add form to answer tab
