+++ formulaire/frontend/src/core/models/distribution/utils.ts
@@ -59,3 +59,29 @@ export const transformDistributionsToTableData = (
     })
     .sort((a, b) => a.responderName.localeCompare(b.responderName));
 };
+
+export const getLatestDistribution = (distributions: IDistribution[]): IDistribution => {
+  return distributions.reduce((latest, current) => {
+    if (!current.dateSending) return latest;
+    if (!latest.dateSending) return current;
+    return new Date(current.dateSending) > new Date(latest.dateSending) ? current : latest;
+  }, distributions[0]);
+};
+
+export const getFirstDistribution = (distributions: IDistribution[]): IDistribution => {
+  return distributions.reduce((first, current) => {
+    if (!current.dateSending) return first;
+    if (!first.dateSending) return current;
+    return new Date(current.dateSending) < new Date(first.dateSending) ? current : first;
+  }, distributions[0]);
+};
+
+export const getNbFinishedDistrib = (distributions: IDistribution[]): number => {
+  return distributions.filter((distribution) => distribution.status === DistributionStatus.FINISHED).length;
+};
+
+export const getFirstDistributionDate = (distributions: IDistribution[]): Date => {
+  const firstDistrib = getFirstDistribution(distributions);
+  return firstDistrib.dateSending ? new Date(firstDistrib.dateSending) : new Date();
+};
+getFirstDistribution(distributions).dateSending
