+++ formulaire/frontend/src/core/models/form/types.ts
@@ -1,5 +1,3 @@
-import { IDistribution } from "../distribution/types";
-
 export interface IForm {
   id: number;
   title: string;
@@ -81,9 +79,4 @@ export interface IFormRight {
 export interface IFormReminderPayload {
   formId: string;
   mail: { link: string; subject: string; body: string };
-}
-
-export interface IFormDistributionsCouple {
-  form: IForm;
-  distributions: IDistribution[];
 }
